refactor: improve packages
