+++ src/main/scala/ulisse/infrastructures/view/map/EditorsView.scala
@@ -1,15 +1,20 @@
 package ulisse.infrastructures.view.map
 
-import ulisse.infrastructures.view.components.container.LayeredContainers.JLayeredPane
+import ulisse.infrastructures.view.components.ExtendedSwing
 import ulisse.infrastructures.view.page.CentralController
 
 import java.awt.BorderLayout
-import scala.swing.{BorderPanel, Component, Dimension}
+import scala.swing.BorderPanel.Position
 import scala.swing.BorderPanel.Position.{Center, East}
+import scala.swing.{BorderPanel, Component, Dimension}
 
-final case class EditorsView() extends JLayeredPane:
+final case class EditorsView():
+  private val mainPane = new ExtendedSwing.LayeredPanel()
+
   private val mapPark       = MapPanel.empty()
+  private val menuPanel     = ExtendedSwing.JBorderPanelItem()
   private val mapController = CentralController.createMap()
 
-  mainPane.peer.add(mapPark.peer)
-  glassPane.peer.add(mapController.component.peer, BorderLayout.EAST)
+  menuPanel.layout(mapController.component) = Position.East
+  mainPane.add(mapPark)
+  mainPane.add(menuPanel)
