+++ src/test/scala/ulisse/infrastructures/view/components/JItemTest.scala
@@ -2,8 +2,7 @@ package ulisse.infrastructures.view.components
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.must.Matchers
-import ulisse.infrastructures.view.components.ui.ExtendedSwing
-import ulisse.infrastructures.view.components.ui.decorators.Styles.defaultRect
+import ulisse.infrastructures.view.components.styles.Styles.defaultRect
 
 class JItemTest extends AnyFlatSpec with Matchers:
   val text: String = "Hello"
