+++ src/test/scala/ulisse/infrastructures/view/components/images/SourceImageTest.scala
@@ -2,7 +2,7 @@ package ulisse.infrastructures.view.components.images
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.must.Matchers
-import ulisse.infrastructures.view.components.ui.decorators.Images.SourceImage
+import ulisse.infrastructures.view.components.styles.Images.SourceImage
 
 object SourceImageTest:
   val rightPath = "icons/route.svg"
