+++ src/test/scala/ulisse/infrastructures/view/components/styles/RectStyleTest.scala
@@ -2,7 +2,7 @@ package ulisse.infrastructures.view.components.styles
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.must.Matchers
-import ulisse.infrastructures.view.components.ui.decorators.Styles.*
+import Styles.*
 
 class RectStyleTest extends AnyFlatSpec with Matchers:
